Ack work after the job has been destroyed
There is a race condition where a job may be acked and then not destroyed (e.g. if the process is killed).
Flip this around so we destroy the job first. We need a plugin to ack the job after a perform in order
to do it at the right time.