Fix inadvertent behaviour in work_off
So, this was fun...

The TomQueue::DelayedJob::Job.reserve method returns nil when an external consumer handled message is found. This was causing the Delayed::Worker.new.work_off(2) calls in the spec to exit the loop after the first message, never attempting to pop a second message, and never failing...

The new TomQueue::Worker::Invoke layer returns true (to indicate a successful job), which meant that TomQueue::Worker.new.work_off(2) _was_ attempting to pop a second message and was therefore timing out as there was no message to pop.

I'm counting this as a bug in the original implementation, as it means that working off a queue of messages would cease after the first external consumer message which was handled.