Allow publishers to be plugged into TomQueue
To implement custom publishing strategies (e.g. moving publishing
to a separate thread), we need to be able to intercept the calls
to publish to an exchange.

This introduces a TomQueue::Publisher class which passes the message
on directly to Bunny. However you can define your own publisher and
replace it with TomQueue.publisher_class = MyClass.