Boot QueueManager if TomQueue.bunny is nil
Let's the QueueManager be created if TomQueue.bunny is nil, which could
be the case if RabbitMQ was unavailable.

The QueueManager will set up the exchange and queues if bunny is
available, otherwise it will try again before setting up the
consumer.

If you are using the consumers, you need to ensure TomQueue.bunny is
set up first.