Honour original priority for deferred jobs
If a job is scheduled for time in the future it will be moved to the
deferred queue and then published back to the Exchange when it is time
to be run.

Unfortunatly when publishing to the Exchange the original
priority of the job isn't honoured resulting in any deferred job being
ending up in the "normal" priority work queue.

This happens because QueueManager.publish expects the priority to be
defined in the "priority" header field however the `publish_*` methods
set "job_priority" instead. Since the "job_priority" header isn't
referenced anywhere else apart from thees functions we change this to
"priority" to make things work as expected.