Only pop messages synchronously
When reserving a job, first we attempt to synchronously pop a message
from each of the queues in priority order until a message is found. If
no message is found, we would then fallback into the wait_for_message
method. This would then set up a subscriber thread on each queue, block
until a message is received and then shut down all the subscribers.

For some reason, if TimeoutErrors are raised when pulling jobs from the
queues and the queue has been emptied, the consumer threads will never
receive any new messages added to the queue and the worker gets stuck.

We can avoid these issues entirely by only pulling messages from the
queue synchronously in a loop. This removes the wait_for_message method
and adds a loop with a sleep(1) into the sync_poll_queues method, which
should have the same effect.