Cancel consumers if subscribe fails
If one of the subscribe calls fails any consumers that were successfully
created beforehand will not be cancelled.

Once one of those uncancelled consumers picks up a message, it will not
be processed because it is stored in an instance variable and that
method invocation has been completed on the main thread.

Additionally because of the global prefetch of 1 on the channel since
the message is not acked it will block any other messages from being
processed.