Introduce forked workers
This commit makes use of our new WorkerSupervisor to fork workers.

Right now we're just wanting to create a single forked worker, but going
forward we can increase this.

This will mean we'll be loading the application once in the before_fork
hook, instead of loading the application for each worker we sping up
(which happens if we call rake jobs:work multiple times)